added support for atlas

+++ engine/src/texture_atlas.rs
@@ -11,15 +11,13 @@ impl TextureAtlas {
         Self { texture, columns, rows }
     }
 
-    pub fn draw(&self, index:u32, x:f32, y:f32) {
+    pub fn draw(&self, index:u16, x:f32, y:f32) {
         let sw = self.texture.width() / self.columns as f32;
         let sh = self.texture.height() / self.rows as f32;
-        let sx = index % self.columns as u32;
-        let sy = index / self.columns as u32;
+        let sx = index % self.columns as u16;
+        let sy = index / self.columns as u16;
         let sx = sx as f32 / self.columns as f32 * self.texture.width();
         let sy = sy as f32 / self.rows as f32 * self.texture.height();
-
-       
 
         let src = Rect::new(sx, sy, sw, sh);
         draw_texture_ex(self.texture, x, y, WHITE, DrawTextureParams {
